# public/openapi.yaml
openapi: 3.0.0
info:
  title: Game API
  version: 1.0.0
  description: API for the interactive game where players can join, shoot at a moving target, and retrieve the current target's position.
servers:
  - url: https://fax-hunt.onrender.com

paths:
  /api/join:
    post:
      summary: Join the game and receive a token.
      operationId: joinGame
      responses:
        "200":
          description: Response with token, username, and color.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JoinResponse"
        "500":
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /api/fire:
    post:
      summary: Shoot at the target.
      operationId: Fire
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FireRequest"
      responses:
        "200":
          description: The shot was successfully processed.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FireResponse"
        "400":
          description: Invalid request due to missing or invalid parameters.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Invalid token.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "429":
          description: Too many requests; the user should retry after some time.
          headers:
            Retry-After:
              description: Number of seconds to wait before retrying.
              schema:
                type: integer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /api/target:
    get:
      summary: Get the current position of the target.
      operationId: getTarget
      responses:
        "200":
          description: Response with the target's coordinates.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TargetResponse"
        "400":
          description: Invalid request due to missing or invalid parameters.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Invalid token.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "429":
          description: Too many requests; the user should retry after some time.
          headers:
            Retry-After:
              description: Number of seconds to wait before retrying.
              schema:
                type: integer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

components:
  schemas:
    JoinResponse:
      type: object
      properties:
        token:
          type: string
          description: Unique token assigned to the player.
        username:
          type: string
          description: Randomly generated username.
        color:
          type: string
          description: Unique color assigned to the player in hexadecimal format.
      required:
        - token
        - username
        - color

    FireRequest:
      type: object
      properties:
        x:
          type: number
          description: X coordinate of the shot.
          minimum: 0
          maximum: 800
        y:
          type: number
          description: Y coordinate of the shot.
          minimum: 0
          maximum: 600
      required:
        - x
        - y

    FireResponse:
      type: object
      properties:
        message:
          type: string
          description: Result message of the shot.
        success:
          type: boolean
          description: Indicates if the request was successfully processed.
        hit:
          type: boolean
          description: Indicates if the shot hit the target.
      required:
        - message
        - success
        - hit

    TargetResponse:
      type: object
      properties:
        x:
          type: number
          description: X coordinate of the target.
          minimum: 0
          maximum: 800
        y:
          type: number
          description: Y coordinate of the target.
          minimum: 0
          maximum: 600
      required:
        - x
        - y

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          description: Error message.
        retryAfter:
          type: integer
          description: Number of seconds to wait before retrying (only for 429 errors).
      required:
        - error

  securitySchemes:
    BearerAuth:
      type: apiKey
      in: header
      name: Authorization
      description: "Insert the token you get from the join call. Format: 'Bearer <token>'"